linux-wandboard: Upgrate to a 4.1-2.0.x based fork
This rebases the Wandboard tree on top of FSLC fork. It avoids
duplicated work across vendors and easy the fix backporting when need.

Due the limited interest on the RT kernel and non-activity on
Wandboard maintenance, the RT kernel has been removed but if someone
is willing to maintain it feel free to send an addition for it.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>